Capstone Web App
This is the entire frontend and first layer of the backend. You can use npm install to install the dependencies, but I also included those under node_modules in case you want to run it locally.

You run the node server and then go to localhost:3000 and everything should work.